# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Athlete {
  id: number;
  firstname: string;
  lastname: string;
  profile: string;
  profile_medium?: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  follower_count?: number | null;
  /** @nullable */
  friend_count?: number | null;
  measurement_preference?: string;
}

export interface ActivityMap {
  id?: string;
  /** @nullable */
  summary_polyline?: string | null;
  /** @nullable */
  polyline?: string | null;
}

export interface Activity {
  id: number;
  name: string;
  type: string;
  sport_type: string;
  start_date: string;
  start_date_local?: string;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  total_elevation_gain?: number;
  average_speed?: number;
  max_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  max_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
  /** @nullable */
  kilojoules?: number | null;
  /** @nullable */
  suffer_score?: number | null;
  /** @nullable */
  kudos_count?: number | null;
  /** @nullable */
  achievement_count?: number | null;
  map?: ActivityMap;
}

export interface Split {
  split: number;
  distance: number;
  elapsed_time: number;
  moving_time: number;
  elevation_difference?: number;
  average_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  average_grade_adjusted_speed?: number | null;
}

export interface Lap {
  id: number;
  lap_index: number;
  name?: string;
  distance: number;
  elapsed_time: number;
  moving_time: number;
  total_elevation_gain?: number;
  average_speed?: number;
  max_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
}

export interface SegmentEffort {
  id: number;
  name: string;
  elapsed_time: number;
  moving_time: number;
  distance?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
  /** @nullable */
  pr_rank?: number | null;
  /** @nullable */
  kom_rank?: number | null;
}

export interface ActivityDetail {
  id: number;
  name: string;
  type: string;
  sport_type: string;
  /** @nullable */
  description?: string | null;
  start_date: string;
  start_date_local?: string;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  total_elevation_gain?: number;
  average_speed?: number;
  max_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  max_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
  /** @nullable */
  max_watts?: number | null;
  /** @nullable */
  weighted_average_watts?: number | null;
  /** @nullable */
  kilojoules?: number | null;
  /** @nullable */
  suffer_score?: number | null;
  /** @nullable */
  kudos_count?: number | null;
  /** @nullable */
  achievement_count?: number | null;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  average_cadence?: number | null;
  map?: ActivityMap;
  splits_metric?: Split[];
  laps?: Lap[];
  segment_efforts?: SegmentEffort[];
}

export interface ActivityStreams {
  time?: number[];
  distance?: number[];
  latlng?: number[][];
  altitude?: number[];
  heartrate?: number[];
  cadence?: number[];
  watts?: number[];
  velocity_smooth?: number[];
  grade_smooth?: number[];
}

export interface ActivityTotals {
  count: number;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  elevation_gain: number;
  /** @nullable */
  achievement_count?: number | null;
}

export interface AthleteStats {
  recent_run_totals: ActivityTotals;
  all_run_totals: ActivityTotals;
  ytd_run_totals: ActivityTotals;
  recent_ride_totals: ActivityTotals;
  all_ride_totals: ActivityTotals;
  ytd_ride_totals: ActivityTotals;
  recent_swim_totals?: ActivityTotals;
  all_swim_totals?: ActivityTotals;
  ytd_swim_totals?: ActivityTotals;
}

export interface WeeklyTotal {
  week_start: string;
  distance: number;
  moving_time: number;
  elevation_gain: number;
  count: number;
  sport_type?: string;
}

export interface SportTypeCount {
  sport_type: string;
  count: number;
  distance: number;
  moving_time: number;
}

export type ListActivitiesParams = {
page?: number;
per_page?: number;
/**
 * @nullable
 */
type?: string | null;
/**
 * @nullable
 */
before?: number | null;
/**
 * @nullable
 */
after?: number | null;
};

export type GetWeeklyStatsParams = {
weeks?: number;
};