# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityMap } from './activityMap';

export interface Activity {
  id: number;
  name: string;
  type: string;
  sport_type: string;
  start_date: string;
  start_date_local?: string;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  total_elevation_gain?: number;
  average_speed?: number;
  max_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  max_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
  /** @nullable */
  kilojoules?: number | null;
  /** @nullable */
  suffer_score?: number | null;
  /** @nullable */
  kudos_count?: number | null;
  /** @nullable */
  achievement_count?: number | null;
  map?: ActivityMap;
}