# lib/api-zod/src/generated/types/activityDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityMap } from './activityMap';
import type { Lap } from './lap';
import type { SegmentEffort } from './segmentEffort';
import type { Split } from './split';

export interface ActivityDetail {
  id: number;
  name: string;
  type: string;
  sport_type: string;
  /** @nullable */
  description?: string | null;
  start_date: string;
  start_date_local?: string;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  total_elevation_gain?: number;
  average_speed?: number;
  max_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  max_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
  /** @nullable */
  max_watts?: number | null;
  /** @nullable */
  weighted_average_watts?: number | null;
  /** @nullable */
  kilojoules?: number | null;
  /** @nullable */
  suffer_score?: number | null;
  /** @nullable */
  kudos_count?: number | null;
  /** @nullable */
  achievement_count?: number | null;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  average_cadence?: number | null;
  map?: ActivityMap;
  splits_metric?: Split[];
  laps?: Lap[];
  segment_efforts?: SegmentEffort[];
}