# lib/api-zod/src/generated/types/activityMap.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityMap {
  id?: string;
  /** @nullable */
  summary_polyline?: string | null;
  /** @nullable */
  polyline?: string | null;
}