# lib/api-zod/src/generated/types/activityStreams.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityStreams {
  time?: number[];
  distance?: number[];
  latlng?: number[][];
  altitude?: number[];
  heartrate?: number[];
  cadence?: number[];
  watts?: number[];
  velocity_smooth?: number[];
  grade_smooth?: number[];
}