# lib/api-zod/src/generated/types/activityTotals.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityTotals {
  count: number;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  elevation_gain: number;
  /** @nullable */
  achievement_count?: number | null;
}