# lib/api-zod/src/generated/types/athlete.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Athlete {
  id: number;
  firstname: string;
  lastname: string;
  profile: string;
  profile_medium?: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  follower_count?: number | null;
  /** @nullable */
  friend_count?: number | null;
  measurement_preference?: string;
}