# lib/api-zod/src/generated/types/athleteStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityTotals } from './activityTotals';

export interface AthleteStats {
  recent_run_totals: ActivityTotals;
  all_run_totals: ActivityTotals;
  ytd_run_totals: ActivityTotals;
  recent_ride_totals: ActivityTotals;
  all_ride_totals: ActivityTotals;
  ytd_ride_totals: ActivityTotals;
  recent_swim_totals?: ActivityTotals;
  all_swim_totals?: ActivityTotals;
  ytd_swim_totals?: ActivityTotals;
}