# lib/api-zod/src/generated/types/lap.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Lap {
  id: number;
  lap_index: number;
  name?: string;
  distance: number;
  elapsed_time: number;
  moving_time: number;
  total_elevation_gain?: number;
  average_speed?: number;
  max_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
}