# lib/api-zod/src/generated/types/segmentEffort.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SegmentEffort {
  id: number;
  name: string;
  elapsed_time: number;
  moving_time: number;
  distance?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  average_watts?: number | null;
  /** @nullable */
  pr_rank?: number | null;
  /** @nullable */
  kom_rank?: number | null;
}