# lib/api-zod/src/generated/types/split.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Split {
  split: number;
  distance: number;
  elapsed_time: number;
  moving_time: number;
  elevation_difference?: number;
  average_speed?: number;
  /** @nullable */
  average_heartrate?: number | null;
  /** @nullable */
  average_grade_adjusted_speed?: number | null;
}