# lib/api-zod/src/generated/types/sportTypeCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SportTypeCount {
  sport_type: string;
  count: number;
  distance: number;
  moving_time: number;
}