# lib/api-zod/src/generated/types/weeklyTotal.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Training Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyTotal {
  week_start: string;
  distance: number;
  moving_time: number;
  elevation_gain: number;
  count: number;
  sport_type?: string;
}